Add files via upload
GUI LOGIN/SIGNUP PANEL
•	AIM: login and signup in a login panel using GUI (Graphical user interface) and sending the email with attachments.
•	Module used:
Tkinter ,smtplib, os,csv,email,filedialog,email.mime.multipart,email.mime.base,email.mime.text
•	Procedure:
•	Import Required Module:
 

•	Storing the credentials 
We have stored the path to the login credentials so that we can further access it and authenticate the user so as to further send or receive the email with or without attachments.

•	Creating the GUI Login/signup page 
This will be displayed to you when your program will run. The main page of the GUI panel is shown below: 
 

In the above main page we used 3 different widgets, i.e., Label, Entry and Button. In the creating of this page we will first create a window manager in which we will add the above widgets.
•	creating a login function:

 

Since, for button widget we have to pass a command and that command is a function so we have to create functions for login, signup and delete user as CheckLogin, Signup and DelUser respectively. 
•	Checklogin:    
1] This function is created to authenticate the user  
2] opening the credentials in a read mode and then checking each record to verify the authorized user.
 

If credentials got matched window will show the message “Login successful” on panel .and at the end send() function is called.
 

If credentials did not get matched window will show the ‘error message” on panel.

•	creating profile window:
In this next step , we create the window to set up two buttons one to compose a new email and other one to logout from your profile and the window created by the function send() will look like.
 


•	Creating window to compose email  
In above code, we have created a new window consisting of two buttons one to compose message and other to logout .Both the buttons will call a function when they are clicked separately those are mail and logout respectively. So the next step is to create a mail() and logout() functions which will open up a new window to compose an email with or without attachment and we will first create the mail() function.
1]creating mail function:
       Below code will generate a window , which looks like:
 
                     
In above code ,attach and send two button widgets are used ,which will call the attach and sendmail function respectively.
2]Creating attach function:
 

This will open all the files from your pc to be selected.
* 
















3]Creating  function to send mail:
In this function,we sent the mail using smtp server
 
If email is sent successfully ,it will show the below panel:
 

•	logout page: 
Now our next step is to create a function to logout using the logout() function .This will display the below image: 

 
•	signup function:
This function will be called when you will click on the signup button. So here is how you will create a function to sign up and the following window will be opened after the button clicked. 
 
 

•	Deleting the credentials for all users :
We can delete the complete records of the users on clicking the delete button. 

 
•	Condition to check and Login/Signup :
Now we have done creating all the functions to validate or to create a user or to delete the credentials. Now the final step is to validate the credentials and accordingly call its first window when the program runs.